+++ app/models/agents/twitter_publish_agent.rb
@@ -25,7 +25,7 @@ module Agents
     end
 
     def working?
-      event_created_within?(options['expected_update_period_in_days']) && most_recent_event.payload['success'] == true && !recent_error_logs?
+      event_created_within?(options['expected_update_period_in_days']) && most_recent_event && most_recent_event.payload['success'] == true && !recent_error_logs?
     end
 
     def default_options
@@ -43,10 +43,11 @@ module Agents
       incoming_events.each do |event|
         tweet_text = Utils.value_at(event.payload, options['message_path'])
         begin
-          publish_tweet tweet_text
+          tweet = publish_tweet tweet_text
           create_event :payload => {
             'success' => true,
             'published_tweet' => tweet_text,
+            'tweet_id' => tweet.id,
             'agent_id' => event.agent_id,
             'event_id' => event.id
           }