+++ app/models/agents/website_agent.rb
@@ -16,11 +16,13 @@ module Agents
 
       Specify a `url` and select a `mode` for when to create Events based on the scraped data, either `all` or `on_change`.
 
+      `url` can be a single url, or an array of urls (for example, for multiple pages with the exact same structure but different content to scrape)
+
       The `type` value can be `xml`, `html`, or `json`.
 
       To tell the Agent how to parse the content, specify `extract` as a hash with keys naming the extractions and values of hashes.
 
-      When parsing HTML or XML, these sub-hashes specify how to extract with a `css` CSS selector and either `'text': true` or `attr` pointing to an attribute name to grab.  An example:
+      When parsing HTML or XML, these sub-hashes specify how to extract with either a `css` CSS selector or a `xpath` XPath expression and either `'text': true` or `attr` pointing to an attribute name to grab.  An example:
 
           'extract': {
             'url': { 'css': "#comic img", 'attr': "src" },
@@ -42,6 +44,8 @@ module Agents
       Set `expected_update_period_in_days` to the maximum amount of time that you'd expect to pass between Events being created by this Agent.  This is only used to set the "working" status.
 
       Set `uniqueness_look_back` to limit the number of events checked for uniqueness (typically for performance).  This defaults to the larger of #{UNIQUENESS_LOOK_BACK} or #{UNIQUENESS_FACTOR}x the number of detected received results.
+
+      Set `force_encoding` to an encoding name if the website does not return a Content-Type header with a proper charset.
     MD
 
     event_description do
@@ -59,8 +63,9 @@ module Agents
           'type' => "html",
           'mode' => "on_change",
           'extract' => {
-            'url' => {'css' => "#comic img", 'attr' => "src"},
-            'title' => {'css' => "#comic img", 'attr' => "title"}
+            'url' => { 'css' => "#comic img", 'attr' => "src" },
+            'title' => { 'css' => "#comic img", 'attr' => "alt" },
+            'hovertext' => { 'css' => "#comic img", 'attr' => "title" }
           }
       }
     end
@@ -84,6 +89,19 @@ module Agents
       if options['uniqueness_look_back'].present?
         errors.add(:base, "Invalid uniqueness_look_back format") unless is_positive_integer?(options['uniqueness_look_back'])
       end
+
+      if (encoding = options['force_encoding']).present?
+        case encoding
+        when String
+          begin
+            Encoding.find(encoding)
+          rescue ArgumentError
+            errors.add(:base, "Unknown encoding: #{encoding.inspect}")
+          end
+        else
+          errors.add(:base, "force_encoding must be a string")
+        end
+      end
     end
 
     def check
@@ -91,66 +109,97 @@ module Agents
       log "Fetching #{options['url']}"
       request_opts = { :followlocation => true }
       request_opts[:userpwd] = options['basic_auth'] if options['basic_auth'].present?
-      request = Typhoeus::Request.new(options['url'], request_opts)
 
-      request.on_failure do |response|
-        error "Failed: #{response.inspect}"
+      requests = []
+
+      if options['url'].kind_of?(Array)
+        options['url'].each do |url|
+           requests.push(Typhoeus::Request.new(url, request_opts))
+        end
+      else
+        requests.push(Typhoeus::Request.new(options['url'], request_opts))
       end
 
-      request.on_success do |response|
-        doc = parse(response.body)
+      requests.each do |request|
+        request.on_failure do |response|
+          error "Failed: #{response.inspect}"
+        end
 
-        if extract_full_json?
-          if store_payload!(previous_payloads(1), doc)
-            log "Storing new result for '#{name}': #{doc.inspect}"
-            create_event :payload => doc
+        request.on_success do |response|
+          body = response.body
+          if (encoding = options['force_encoding']).present?
+            body = body.encode(Encoding::UTF_8, encoding)
           end
-        else
-          output = {}
-          options['extract'].each do |name, extraction_details|
-            result = if extraction_type == "json"
-                       output[name] = Utils.values_at(doc, extraction_details['path'])
-                     else
-                       output[name] = doc.css(extraction_details['css']).map { |node|
-                         if extraction_details['attr']
-                           node.attr(extraction_details['attr'])
-                         elsif extraction_details['text']
-                           node.text()
-                         else
-                           error "'attr' or 'text' is required on HTML or XML extraction patterns"
-                           return
-                         end
-                       }
-                     end
-            log "Extracting #{extraction_type} at #{extraction_details['path'] || extraction_details['css']}: #{result}"
-          end
+          doc = parse(body)
 
-          num_unique_lengths = options['extract'].keys.map { |name| output[name].length }.uniq
-
-          if num_unique_lengths.length != 1
-            error "Got an uneven number of matches for #{options['name']}: #{options['extract'].inspect}"
-            return
-          end
-      
-          old_events = previous_payloads num_unique_lengths.first
-          num_unique_lengths.first.times do |index|
-            result = {}
-            options['extract'].keys.each do |name|
-              result[name] = output[name][index]
-              if name.to_s == 'url'
-                result[name] = URI.join(options['url'], result[name]).to_s if (result[name] =~ URI::DEFAULT_PARSER.regexp[:ABS_URI]).nil?
+          if extract_full_json?
+            if store_payload!(previous_payloads(1), doc)
+              log "Storing new result for '#{name}': #{doc.inspect}"
+              create_event :payload => doc
+            end
+          else
+            output = {}
+            options['extract'].each do |name, extraction_details|
+              if extraction_type == "json"
+                result = Utils.values_at(doc, extraction_details['path'])
+                log "Extracting #{extraction_type} at #{extraction_details['path']}: #{result}"
+              else
+                case
+                when css = extraction_details['css']
+                  nodes = doc.css(css)
+                when xpath = extraction_details['xpath']
+                  nodes = doc.xpath(xpath)
+                else
+                  error "'css' or 'xpath' is required for HTML or XML extraction"
+                  return
+                end
+                unless Nokogiri::XML::NodeSet === nodes
+                  error "The result of HTML/XML extraction was not a NodeSet"
+                  return
+                end
+                result = nodes.map { |node|
+                  if extraction_details['attr']
+                    node.attr(extraction_details['attr'])
+                  elsif extraction_details['text']
+                    node.text()
+                  else
+                    error "'attr' or 'text' is required on HTML or XML extraction patterns"
+                    return
+                  end
+                }
+                log "Extracting #{extraction_type} at #{xpath || css}: #{result}"
               end
+              output[name] = result
             end
 
-            if store_payload!(old_events, result)
-              log "Storing new parsed result for '#{name}': #{result.inspect}"
-              create_event :payload => result
+            num_unique_lengths = options['extract'].keys.map { |name| output[name].length }.uniq
+
+            if num_unique_lengths.length != 1
+              error "Got an uneven number of matches for #{options['name']}: #{options['extract'].inspect}"
+              return
+            end
+        
+            old_events = previous_payloads num_unique_lengths.first
+            num_unique_lengths.first.times do |index|
+              result = {}
+              options['extract'].keys.each do |name|
+                result[name] = output[name][index]
+                if name.to_s == 'url'
+                  result[name] = URI.join(options['url'], result[name]).to_s if (result[name] =~ URI::DEFAULT_PARSER.regexp[:ABS_URI]).nil?
+                end
+              end
+
+              if store_payload!(old_events, result)
+                log "Storing new parsed result for '#{name}': #{result.inspect}"
+                create_event :payload => result
+              end
             end
           end
         end
+
+        hydra.queue request
+        hydra.run
       end
-      hydra.queue request
-      hydra.run
     end
 
     private